Move ionic_levels to atomic.py
